prove main end-to-end lemma for Viper correctness with pre- and postcondition (no support for old expressions)
The Viper correctness statement takes old states into account. However, the end-to-end lemma uses a state relation that does not track the old state. So the premises of the lemma can only be satisfied if there are no old expressions. Once we support old expressions we will have to adjust the state relation and as a result this lemma.