Turn state accessor and update definitions into functions, which simplifies proofs.
Since these functions are defined via record getters and record updates, the Isabelle simplifier already has all the lemmas needed to prove basic properties involving these functions. Thus, it makes sense to not hide the functions via definitions and instead to use functions.

The lemmas proving the basic properties are not needed anymore (with minor exceptions such as needing to instantiate parameters in a goal via a rule application of such a lemma).