Removed remaining sorry in ViperBoogieRelUtil, which was achieved by adjusting the state relation.
There is now an option that controls whether the state consistency should be enforced, which allows one to derive the state relation w.r.t. a consistent Viper state via intermediate "helper" Viper states that may not be consistent (for example, updating the heap and mask in separate steps).
Another minor change is that heap_var_rel and mask_var_rel directly take the field translation as a parameter instead of the entire translation record.